backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.7.2
python-dotenv==1.0.1
pypdf==4.3.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
numpy
faiss-cpu
